teste-api/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "GabrielLima2803", email = "gabriellima2803@gmail.com"},
]
dependencies = ["django>=5.1.7", "djangorestframework>=3.16.0", "django-cors-headers>=4.7.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
